requirements.txt
arrow
beautifulsoup4
bibtexparser
PySocks
requests[security]
requests[socks]
stem
fake_useragent
selenium
python-dotenv
free-proxy